Header - moved state management logic to a separate hook Added types structure
