+++ client/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './transaction.types';
+export * from './category.types';
+export * from './goal.types';
